daemon-per-session (#4)
This is a ground-up rewrite of this tool.  We have dramatically simplified how this tool works.  Instead of having a single daemon that manages all sessions, we have a daemon-per-session.

We employ a simple architecture: `fork() -> setsid() on child -> forkpty() -> attach on parent`

This has a bunch of benefits, one of which is we can safely remove libxev and use simple `poll()` loops for daemon and clients.