fix: dont use alt screen on attach
The problem is that the client switches to the alternate screen buffer on attach, which doesn't have scrollback. When ghostty replays content, it goes into the alt screen where scrollback doesn't accumulate.

The alternate screen buffer (\x1b[?1049h) never has scrollback - that's by design in terminals.

To get scrollback working properly we now stay on main screen and clear on detach.