refactor: when daemon is unresponsive, remove it
This commit does a better job removing a unix socket file when
the daemon is unresponsive, likely indicating that the process has
already been killed.

We accomplish this by creating a probSession fn that will connect to the
unix socket, send the Info event, and timeout after 1 second.  After
that point we remove the unix socket file.