refactor: remove ctrl+b + d detach shortcut
This feature was annoying to implement because we can't simply check for
a single byte, we first have to detect if ctrl+b was pressed then check
for d.

In the end I don't like the complexity it introduces and it seems pretty low
value since we should recommend users simply close their terminal
window.